requirements.txt
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
opencv-python==4.7.0.72
pynput==1.7.3
scipy==1.10.0
sounddevice==0.4.3
